# project/openapi.yaml
openapi: 3.0.3
info:
  title: 我的数字老师 - 后端接口
  version: 0.1.0
  description: |
    由当前 uni-app 前端项目整理出的"需要/预留"的后端接口规格。
    可直接导入 Swagger / Apifox / Postman。
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        ok:
          type: boolean
        code:
          type: integer
          example: 0
        message:
          type: string
          example: "ok"
        data:
          nullable: true
...
    ChatAttachment:
      type: object
      required: [type, url]
      properties:
        type:
          type: string
          enum: [image, file]
        url:
          type: string
    ChatRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/ChatAttachment'
        meta:
          type: object
          additionalProperties: true
    ChatResponseData:
      type: object
      properties:
        answer:
          type: string
        conversationId:
          type: string
        messageId:
          type: string
    UploadResponseData:
      type: object
      properties:
        url:
          type: string
        mime:
          type: string
        size:
          type: integer
...
paths:
...
  /chat:
    post:
      tags: [Chat]
      summary: 发送对话
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ChatResponseData'
  /upload/image:
    post:
      tags: [Upload]
      summary: 上传图片（聊天附件）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UploadResponseData'
  /upload/file:
    post:
      tags: [Upload]
      summary: 上传文件（聊天附件）
      security:
        - bearerAuth: []
  /upload/voice:
    post:
      tags: [Upload]
      summary: Upload voice recording
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /voice/chat:
    post:
      tags: [Voice]
      summary: Voice understanding chat (ASR)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [voiceUrl]
              properties:
                voiceUrl:
                  type: string
                  example: /files/voice/xxx.aac
                conversationId:
                  type: string
                  nullable: true
                meta:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UploadResponseData'
...
  /avatar/upload:
    post:
      tags: [Avatar]
      summary: 上传头像照片（数字人采集）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UploadResponseData'
...